2-data-modeling/requirements.txt
cassandra-driver
pandas
